+++ src/Parser.php
@@ -82,7 +82,7 @@ class Parser extends BaseObject
      * Arguments is not required to be variadic `...$children`. It would be much better
      * to declare exact amount of arguments with respect to corresponding rule(s).
      *
-     * Return value of a callback (unless it's `null`) will be use in `make()` method
+     * Return value of a callback (unless it's `null`) will be used in `make()` method
      * on a node. Callback itself should to use children nodes' `made()` values to
      * evaluate the result. To apply `null` value to a node you need to call `make(null)`
      * manually in action callback, but it is not necessary since default `made()` value is `null`.