+++ CHANGELOG.md
@@ -4,8 +4,11 @@ LR(0) parser Change Log
 1.3.2
 -----
 
+*   Deprecated: class `\VovanVE\parser\LexerBuilder`. Use `use VovanVE\parser\lexer\Lexer`
+    constructor directly. Configuration methods of `LexerBuilder` introduced in `Lexer` to
+    simplify migration, but remember that `Lexer` is immutable.
 *   Add inline quoted tokens in grammar source like `Sum(add): Sum "+" Product`.
-*   Add ability to omit some useless tokens from the resulting tree.
+*   Add ability to omit some useless tokens from the resulting tree like `A: foo .bar baz`.
 *   Add action shortcut to bubble up the only child's `made()` value.
 *   Enh: Returning `null` from action handler will no longer change node's `made()` value.